refactor(monoid): pcm --- partial commutative monoids
Initially, we formalized partial monoids without the assumption of
commutativity. We believed that synchronization algebras are not
obligated to be commutativity. It seems, however, that
the commutativity can be assumed for synchonization algebras
and thus we can inherit SA from PCM.

The new idea for SA is to split operations into call/return sites
(a.k.a. send/receive, request/response etc).
For example for read/write SA we will have:

L := W_req(x,v)
   | W_rsp(x,v)
   | R_req(x)
   | R_rsp(x,v)

Then we can define SA as follows.
  W_rsp(x,v) + R_req(x) = R_req(x) + W_rsp(x,v) = R_rsp(x,v)