docs(lts): note on trace monoid
Unfortunatelly, it is impossible to declare
partial monoid instance for traces (in the current formulation).
In order to do that we need to embed an absorbing element
into trace type, but then we will lose `subType` injection to `seq`,
we will have to perform case-analysis everywhere,
and we will face other obstacles.
Thus, a better solution would be to give up
on monoid instance for traces,
and declare it (later) only for trace languages.