feat(pomset): STUCK with pomsets as quotient types
It seems that the idea of implementing pomsets as mathcomp's quotients
is infeasible currently. The reason is that required equivalence relation
on finite lposets (isomoprhism) is heteregeneous, it can
relate lposets with different carriers.
Indeed, we need pomset languages to be defined as
decidable predicates over pomsets with possibly
different carriers (e.g. a pomset language
corresponding to an automaton should contain
all pomsets of size 1, 2, ..., N related
to partial runs of the automaton).
However, Mathcomp's quotients currently require homogeneous equivalence relation.

Thus we need to abandon this encoding for now.
A solutions to the problem could be.

(1) Generalize mathcomp's quotients to work with heterogeneous
relations (is that even possible?). But event if it could be done,
it would require a lot of work and communication with mathcomp team
and thus cannot be achieved in the near future.

(2) Alternatively, we can work with finitely supported lposets
and pomsets. In this setting we would avoid the problem of
the heterogeneous equivalence, since the lposets would
have the same carrier, but different domain.
This can also work nicely with the operational semantics
which constructs lposets & event structures incrementally,
since there we also work with finitely supported lposets.
Thus we should try this approach in combination with op.sem.
theory of lposets & event structures.