Pomsets as quotient types (#135)
This PR features an implementation of a pomset as a quotient type over finitely supported labeled posets.

For doing mathcomp-style quotients we need the following:
- `T : choiceType`
- `e : rel T` --- decidable homogeneous equivalence relation. 

The idea of implementing pomsets as quotients over finite posets (i.e. `lFinPoset.class`) does not work. 
The reason is that required equivalence relation on finite lposets (isomoprhism) is heteregeneous, it can
relate lposets with different carriers. Indeed, we need pomset languages to be defined as
decidable predicates over pomsets with possibly different carriers (e.g. a pomset language
corresponding to an automaton should contain all pomsets of size 1, 2, ..., N related to partial runs of the automaton).
However, mathcomp's quotients require homogeneous equivalence relation.

Thus, we take alternative approach and work with finitely supported lposets and pomsets. 
In this setting we would avoid the problem of the heterogeneous equivalence, since the lposets would 
have the same carrier, but different support.
This should also work nicely with the operational semantics 
which constructs lposets & event structures incrementally,
since there we also work with finitely supported lposets.